NMS-16486: Extracted caches from DocumentEnricherImpl
