+++ features/telemetry/protocols/common/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -19,4 +19,79 @@
             <entry key="registration.export" value="true" />
         </service-properties>
     </service>
+
+    <!-- Configuration properties -->
+    <cm:property-placeholder id="cacheConfigProperties" persistent-id="org.opennms.features.telemetry.protocols.common" update-strategy="reload">
+        <cm:default-properties>
+            <cm:property name="nodeInfoCache.enabled" value="true" /> <!-- Set to false to disable caching -->
+            <cm:property name="nodeInfoCache.maximumSize" value="10000"/> <!-- Set value for unlimited size -->
+            <cm:property name="nodeInfoCache.expireAfterWrite" value="0"/> <!-- in seconds. Set to 0 to never evict elements -->
+            <cm:property name="nodeInfoCache.recordStats" value="true"/> <!-- Set to false to not expose cache statistics via jmx -->
+
+            <cm:property name="nodeMetadataCache.enabled" value="true" /> <!-- Set to false to disable caching -->
+            <cm:property name="nodeMetadataCache.maximumSize" value="10000"/> <!-- Set value for unlimited size -->
+            <cm:property name="nodeMetadataCache.expireAfterWrite" value="0"/> <!-- in seconds. Set to 0 to never evict elements -->
+            <cm:property name="nodeMetadataCache.recordStats" value="true"/> <!-- Set to false to not expose cache statistics via jmx -->
+        </cm:default-properties>
+    </cm:property-placeholder>
+
+    <!-- Configure node cache -->
+    <bean id="nodeInfoCacheConfig" class="org.opennms.core.cache.CacheConfig">
+        <argument value="nodeInfoCache"/>
+        <property name="metricRegistry" ref="flowCacheMetricRegistry" />
+        <property name="enabled" value="${nodeInfoCache.enabled}" />
+        <property name="maximumSize" value="${nodeInfoCache.maximumSize}" />
+        <property name="expireAfterWrite" value="${nodeInfoCache.expireAfterWrite}" />
+        <property name="recordStats" value="${nodeInfoCache.recordStats}" />
+    </bean>
+
+    <bean id="nodeMetadataCacheConfig" class="org.opennms.core.cache.CacheConfig">
+        <argument value="nodeMetadataCache"/>
+        <property name="metricRegistry" ref="flowCacheMetricRegistry" />
+        <property name="enabled" value="${nodeMetadataCache.enabled}" />
+        <property name="maximumSize" value="${nodeMetadataCache.maximumSize}" />
+        <property name="expireAfterWrite" value="${nodeMetadataCache.expireAfterWrite}" />
+        <property name="recordStats" value="${nodeMetadataCache.recordStats}" />
+    </bean>
+
+    <!-- Configure references -->
+    <reference id="nodeDao" interface="org.opennms.netmgt.dao.api.NodeDao" availability="mandatory" />
+    <reference id="ipInterfaceDao" interface="org.opennms.netmgt.dao.api.IpInterfaceDao" availability="mandatory" />
+    <reference id="interfaceToNodeCache" interface="org.opennms.netmgt.dao.api.InterfaceToNodeCache" availability="mandatory" />
+
+    <!-- Configure NodeMetadataCache -->
+    <bean id="nodeMetadataCache" class="org.opennms.netmgt.telemetry.protocols.common.cache.NodeMetadataCacheImpl">
+        <argument ref="nodeInfoCacheConfig"/>
+        <argument ref="nodeMetadataCacheConfig"/>
+        <argument ref="flowCacheMetricRegistry"/>
+        <argument ref="nodeDao"/>
+        <argument ref="ipInterfaceDao"/>
+        <argument ref="interfaceToNodeCache"/>
+    </bean>
+    <service ref="nodeMetadataCache" interface="org.opennms.netmgt.telemetry.protocols.common.cache.NodeMetadataCache">
+        <service-properties>
+            <entry key="registration.export" value="true" />
+        </service-properties>
+    </service>
+
+    <!-- Configure Metrics -->
+    <bean id="flowCacheMetricRegistry" class="com.codahale.metrics.MetricRegistry"/>
+    <service ref="flowCacheMetricRegistry" interface="com.codahale.metrics.MetricSet">
+        <service-properties>
+            <entry key="name" value="org.opennms.netmgt.telemetry.protocols.common.cache" />
+            <entry key="description" value="Node Caches" />
+        </service-properties>
+    </service>
+    <bean id="flowCacheMetricRegistryJmxReporterBuilder" class="com.codahale.metrics.JmxReporter" factory-method="forRegistry">
+        <argument ref="flowCacheMetricRegistry"/>
+    </bean>
+    <bean id="flowCacheMetricRegistryDomainedJmxReporterBuilder" factory-ref="flowCacheMetricRegistryJmxReporterBuilder" factory-method="inDomain">
+        <argument value="org.opennms.netmgt.flows"/>
+    </bean>
+    <bean id="flowCacheMetricRegistryJmxReporter"
+          factory-ref="flowCacheMetricRegistryDomainedJmxReporterBuilder"
+          factory-method="build"
+          init-method="start"
+          destroy-method="stop" />
+
 </blueprint>
